wmhdplop: fix integer overflow
After a while, the procstats fields will overflow the
signed int field and the hd leds will be permanently on.

Use long instead of int for these counters

Gentoo original bug: https://bugs.gentoo.org/show_bug.cgi?id=325615

+++ wmhdplop/devnames.h
@@ -9,7 +9,7 @@ typedef struct DiskList {
   unsigned major, minor;
   int hd_id, part_id; /* part_id = 0 for disks */
   int enable_hddtemp;
-  int nr, nw, touched_r, touched_w;
+  long nr, nw, touched_r, touched_w;
   struct DiskList *next;
 } DiskList;
